+++ test/token-spec.js
